Renamed hook from useAdmin to useOfficers
