+++ src/pages/index.ts
@@ -1,8 +1,6 @@
-
 export { NotFound } from "./NotFound";
 export { Profile } from "./Profile/Profile";
 export { Events } from "./Events/Events";
-
 export { Home } from "./Home/Home";
-export { Officers } from "./Officers";
+export { Officers } from "./Officers/Officers";
 export { Projects } from "./Projects/Projects";
