+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/assets/dynamicmomo.tsx","./src/assets/stars.tsx","./src/assets/animations/joinmobitoday.tsx","./src/assets/animations/rotatingtext.tsx","./src/assets/animations/scrollreveal.tsx","./src/assets/animations/splittext.tsx","./src/assets/background/galaxy.tsx","./src/assets/background/galaxybg.tsx","./src/assets/momo/momoexpression.tsx","./src/assets/momo/momoface.tsx","./src/assets/waves/wave.tsx","./src/assets/waves/waves.ts","./src/components/colorpicker.tsx","./src/components/footer.tsx","./src/components/herosection.tsx","./src/components/layout.tsx","./src/components/logoutbutton.tsx","./src/components/navlinks.tsx","./src/components/navbar.tsx","./src/components/publiclayout.tsx","./src/components/techstackfield.tsx","./src/components/themebutton.tsx","./src/components/ui/reveal.tsx","./src/components/ui/color-mode.tsx","./src/components/ui/password-input.tsx","./src/components/ui/provider.tsx","./src/components/ui/toaster.tsx","./src/components/ui/tooltip.tsx","./src/context/form-context.ts","./src/data/officers.ts","./src/data/projects.ts","./src/data/tech_stack.ts","./src/helpers/format.ts","./src/helpers/project-form.ts","./src/helpers/scroll.tsx","./src/hooks/supabaseclient.ts","./src/hooks/useadmin.ts","./src/hooks/useauth.ts","./src/hooks/useevents.ts","./src/hooks/useprofile.ts","./src/hooks/useprojects.ts","./src/pages/notfound.tsx","./src/pages/officers.tsx","./src/pages/index.ts","./src/pages/auth/forgotpassword.tsx","./src/pages/auth/formfields.tsx","./src/pages/auth/loginform.tsx","./src/pages/auth/registerform.tsx","./src/pages/auth/resetpassword.tsx","./src/pages/events/eventcard.tsx","./src/pages/events/eventdialog.tsx","./src/pages/events/eventlist.tsx","./src/pages/events/eventlistskeleton.tsx","./src/pages/events/events.tsx","./src/pages/events/rsvpbutton.tsx","./src/pages/events/upcomingevents.tsx","./src/pages/home/home.tsx","./src/pages/home/jointoday.tsx","./src/pages/home/officercard.tsx","./src/pages/home/officerdialog.tsx","./src/pages/home/officers.tsx","./src/pages/home/projects.tsx","./src/pages/home/vibesandvalues.tsx","./src/pages/home/whatweoffer.tsx","./src/pages/home/whoarewe.tsx","./src/pages/profile/editprofilemenu.tsx","./src/pages/profile/membershipcard.tsx","./src/pages/profile/profile.tsx","./src/pages/profile/profilepicture.tsx","./src/pages/profile/profileprojects.tsx","./src/pages/profile/sidebar.tsx","./src/pages/profile/sociallinks.tsx","./src/pages/projects/fileuploadinput.tsx","./src/pages/projects/formfields.tsx","./src/pages/projects/projectcard.tsx","./src/pages/projects/projectdialog.tsx","./src/pages/projects/projectform.tsx","./src/pages/projects/projectformprovider.tsx","./src/pages/projects/projectimagepreview.tsx","./src/pages/projects/projects.tsx","./src/pages/projects/buttons/deleteprojectbutton.tsx","./src/pages/projects/buttons/updateprojectbutton.tsx","./src/pages/projects/routes/createprojectpage.tsx","./src/pages/projects/routes/updateprojectpage.tsx","./src/providers/authroutes.tsx","./src/providers/authsessionlistener.tsx","./src/providers/profileprovider.tsx","./src/providers/protectedroutes.tsx","./src/routes/auth.tsx","./src/routes/protected.tsx","./src/routes/public.tsx","./src/schema/projects.ts","./src/styles/extend.ts","./src/types/index.ts"],"errors":true,"version":"5.8.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/assets/dynamicmomo.tsx","./src/assets/stars.tsx","./src/assets/animations/joinmobitoday.tsx","./src/assets/animations/rotatingtext.tsx","./src/assets/animations/scrollreveal.tsx","./src/assets/animations/splittext.tsx","./src/assets/background/galaxy.tsx","./src/assets/background/galaxybg.tsx","./src/assets/momo/momoexpression.tsx","./src/assets/momo/momoface.tsx","./src/assets/waves/wave.tsx","./src/assets/waves/waves.ts","./src/components/colorpicker.tsx","./src/components/footer.tsx","./src/components/herosection.tsx","./src/components/layout.tsx","./src/components/logoutbutton.tsx","./src/components/navlinks.tsx","./src/components/navbar.tsx","./src/components/publiclayout.tsx","./src/components/techstackfield.tsx","./src/components/themebutton.tsx","./src/components/ui/reveal.tsx","./src/components/ui/color-mode.tsx","./src/components/ui/password-input.tsx","./src/components/ui/provider.tsx","./src/components/ui/toaster.tsx","./src/components/ui/tooltip.tsx","./src/context/form-context.ts","./src/data/officers.ts","./src/data/projects.ts","./src/data/tech_stack.ts","./src/helpers/format.ts","./src/helpers/project-form.ts","./src/helpers/scroll.tsx","./src/hooks/supabaseclient.ts","./src/hooks/useauth.ts","./src/hooks/useevents.ts","./src/hooks/useofficer.ts","./src/hooks/useprofile.ts","./src/hooks/useprojects.ts","./src/hooks/usersvp.ts","./src/pages/notfound.tsx","./src/pages/index.ts","./src/pages/auth/forgotpassword.tsx","./src/pages/auth/formfields.tsx","./src/pages/auth/loginform.tsx","./src/pages/auth/registerform.tsx","./src/pages/auth/resetpassword.tsx","./src/pages/events/eventcard.tsx","./src/pages/events/eventdialog.tsx","./src/pages/events/eventlist.tsx","./src/pages/events/eventlistskeleton.tsx","./src/pages/events/events.tsx","./src/pages/events/rsvpbutton.tsx","./src/pages/events/upcomingevents.tsx","./src/pages/home/home.tsx","./src/pages/home/jointoday.tsx","./src/pages/home/projects.tsx","./src/pages/home/vibesandvalues.tsx","./src/pages/home/whatweoffer.tsx","./src/pages/home/whoarewe.tsx","./src/pages/officers/officercard.tsx","./src/pages/officers/officerdialog.tsx","./src/pages/officers/officers.tsx","./src/pages/profile/editprofilemenu.tsx","./src/pages/profile/membershipcard.tsx","./src/pages/profile/profile.tsx","./src/pages/profile/profilepicture.tsx","./src/pages/profile/profileprojects.tsx","./src/pages/profile/sidebar.tsx","./src/pages/profile/sociallinks.tsx","./src/pages/projects/fileuploadinput.tsx","./src/pages/projects/formfields.tsx","./src/pages/projects/projectcard.tsx","./src/pages/projects/projectdialog.tsx","./src/pages/projects/projectform.tsx","./src/pages/projects/projectformprovider.tsx","./src/pages/projects/projectimagepreview.tsx","./src/pages/projects/projects.tsx","./src/pages/projects/buttons/deleteprojectbutton.tsx","./src/pages/projects/buttons/updateprojectbutton.tsx","./src/pages/projects/routes/createprojectpage.tsx","./src/pages/projects/routes/updateprojectpage.tsx","./src/providers/authroutes.tsx","./src/providers/authsessionlistener.tsx","./src/providers/profileprovider.tsx","./src/providers/protectedroutes.tsx","./src/routes/auth.tsx","./src/routes/protected.tsx","./src/routes/public.tsx","./src/schema/projects.ts","./src/styles/extend.ts","./src/types/index.ts"],"version":"5.8.3"}
